Some editings
